+++ python/marvin/tools/__init__.py
@@ -7,3 +7,6 @@ from .modelcube import ModelCube
 from .plate import Plate
 from .rss import RSS, RSSFiber
 from .spaxel import Bin, Spaxel
+from .image import Image
+from .query import Query
+from .results import Results