+++ python/marvin/tools/cube.py
@@ -7,7 +7,7 @@
 # @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)
 #
 # @Last modified by: José Sánchez-Gallego (gallegoj@uw.edu)
-# @Last modified time: 2018-08-04 13:43:28
+# @Last modified time: 2018-07-30 11:33:31
 
 
 from __future__ import absolute_import, division, print_function
@@ -41,8 +41,6 @@ class Cube(MarvinToolsClass, NSAMixIn, GetApertureMixIn):
     See `~.MarvinToolsClass` and `~.NSAMixIn` for a list of input parameters.
 
     """
-
-    _qualflag = 'DRP3QUAL'
 
     def __init__(self, input=None, filename=None, mangaid=None, plateifu=None,
                  mode=None, data=None, release=None,
@@ -270,7 +268,7 @@ class Cube(MarvinToolsClass, NSAMixIn, GetApertureMixIn):
                             np.array(self._wavelength),
                             ivar=self._get_extension_data(name, 'ivar'),
                             mask=self._get_extension_data(name, 'mask'),
-                            unit=model.unit, pixmask_flag=self.header['MASKNAME'])
+                            unit=model.unit, pixmask_flag=model.pixmask_flag)
 
         return datacube
 
@@ -288,7 +286,7 @@ class Cube(MarvinToolsClass, NSAMixIn, GetApertureMixIn):
                             wavelength=np.array(self._wavelength),
                             std=self._get_extension_data(name, 'std'),
                             unit=model.unit,
-                            pixmask_flag=self.header['MASKNAME'])
+                            pixmask_flag=model.pixmask_flag)
 
         return spectrum
 
@@ -497,7 +495,7 @@ class Cube(MarvinToolsClass, NSAMixIn, GetApertureMixIn):
                                                     std=data['std'],
                                                     wavelength=self._wavelength,
                                                     unit=dm.unit,
-                                                    pixmask_flag=self.header['MASKNAME'])
+                                                    pixmask_flag=dm.pixmask_flag)
 
         if self.data_origin == 'api':
 
@@ -527,7 +525,7 @@ class Cube(MarvinToolsClass, NSAMixIn, GetApertureMixIn):
                                                     mask=data[dm.name]['mask'],
                                                     wavelength=data['wavelength'],
                                                     unit=dm.unit,
-                                                    pixmask_flag=self.header['MASKNAME'])
+                                                    pixmask_flag=dm.pixmask_flag)
 
         return cube_quantities