+++ python/marvin/utils/datamodel/drp/MPL.py
@@ -68,40 +68,42 @@ RSS_extensions = [
 MPL4 = DRPCubeDataModel('MPL-4', aliases=['MPL4', 'v1_5_1'],
                         datacubes=MPL4_datacubes,
                         spectra=MPL4_spectra,
-                        bitmasks=get_maskbits('MPL-4'))
+                        bitmasks=get_maskbits('MPL-4'),
+                        qual_flag='DRP3QUAL')
 
 MPL5 = DRPCubeDataModel('MPL-5', aliases=['MPL5', 'v2_0_1'],
                         datacubes=MPL4_datacubes,
                         spectra=MPL4_spectra,
-                        bitmasks=get_maskbits('MPL-5'))
+                        bitmasks=get_maskbits('MPL-5'),
+                        qual_flag='DRP3QUAL')
 
 DR14 = DRPCubeDataModel('DR14', aliases=['DR14', 'v2_1_2'],
                         datacubes=MPL4_datacubes,
                         spectra=MPL4_spectra,
-                        bitmasks=get_maskbits('MPL-5'))
+                        bitmasks=get_maskbits('MPL-5'),
+                        qual_flag='DRP3QUAL')
 
 MPL6 = DRPCubeDataModel('MPL-6', aliases=['MPL6', 'v2_3_1'],
                         datacubes=MPL4_datacubes + MPL6_datacubes,
                         spectra=MPL4_spectra + MPL6_spectra,
-                        bitmasks=get_maskbits('MPL-6'))
+                        bitmasks=get_maskbits('MPL-6'),
+                        qual_flag='DRP3QUAL')
 
 MPL7 = DRPCubeDataModel('MPL-7', aliases=['MPL7', 'v2_4_3', 'DR15'],
                         datacubes=MPL4_datacubes + MPL6_datacubes,
                         spectra=MPL4_spectra + MPL6_spectra,
-                        bitmasks=get_maskbits('MPL-7'))
+                        bitmasks=get_maskbits('MPL-7'),
+                        qual_flag='DRP3QUAL')
 
 # The DRP Cube Datamodel
 datamodel = DRPCubeDataModelList([MPL4, MPL5, MPL6, MPL7, DR14])
 
 # Define the RSS Datamodel. Start by copying the Cube datamodel for convenience.
 datamodel_rss = datamodel.copy()
+
 for release in datamodel_rss:
     datamodel_rss[release] = datamodel_rss[release].to_rss()
 
     flux = datamodel_rss[release].rss.flux
     flux.description = 'Row-stacked spectra from all exposures for the target'
     flux.unit = flux.unit * spaxel_unit / fiber_unit
-
-    # Append specific RSS extensions
-    for ext in RSS_extensions:
-        datamodel_rss[release].rss.append(ext)